Solucao/requirements.txt
plotly
psycopg2-binary
Faker
streamlit
pandas
matplotlib